Login: don't store path for unload requests
When a site is requested but the user not logged in yet,
the requested url is stored and the user will be redirected
to that site after a successful login.
When the users leaves or reloads the page, an unload request is sent.
When he reloads the page while the http session is disposed,
the unload request will be stored and the user redirected to the unload
path after the login, which obviously does not work.

-> Don't store paths for requests that cannot be redirected. We could
explicitly check for unload requests, but a more general logic to detect
such requests already exists. Since the unload request is a POST request
that logic will work fine.

https://www.eclipse.org/forums/index.php/t/1107743/

293660

Change-Id: I8f57a52a691edefa17107182cfe3150018af43ef
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/179828
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>